feat: refactor MLIR dialects (#99)

+++ include/athena/backend/llvm/runtime/LaunchCommand.h
@@ -16,20 +16,22 @@
 #ifndef ATHENA_LAUNCHCOMMAND_H
 #define ATHENA_LAUNCHCOMMAND_H
 
+#include <cstdint>
+
 struct ArgDesc {
   enum ArgType { TENSOR = 0, DATA = 1 };
-  size_t size;
+  uint64_t size;
   void* arg;
   ArgType type;
 };
 
 struct LaunchCommand {
   const char* kernelName;
-  uint32_t argsCount;
+  uint64_t argsCount;
   ArgDesc* args;
-  size_t workDim;
-  size_t* globalSize;
-  size_t* localSize;
+  uint64_t workDim;
+  uint64_t* globalSize;
+  uint64_t* localSize;
 };
 
 #endif // ATHENA_LAUNCHCOMMAND_H
